gateway/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gateway"
version = "0.1.0"
description = "Gateway service for Magi project"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "redis",
    "websockets",
    "python-jose[cryptography]",
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "httpx",
    "pydantic-settings",
    "python-dotenv",
    "async-timeout"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto" 